2nd Commit 04/17/2020
Slope feature

+++ AdvancedCalcMannNguyen6/src/Factorial.java
@@ -1,12 +1,16 @@
 
 public class Factorial {
 	
-	public int factorial(int n) {
+	// finds factorial of n (entered number)
+	public double factorial(double n) {
 			
+		// base case
 		if(n == 1) {
 			System.out.println(n + "! = 1");
 			return 1;
-		} else {
+		}
+		// factorial recursion
+		else {
 			System.out.println(n + "! = "  + n + " * " + (n-1) + "!");
 			return n * factorial(n-1);
 		}
